requirement.txt
Django
djangorestframework
pillow
